update sdk return data type
